+++ R/strat_quantiles.R
@@ -5,19 +5,18 @@
 #'
 #' @inheritParams strat_kmeans
 #' @inheritParams strat_breaks
-#' @param nStrata Numeric. Number of quantiles to stratify \code{mraster}.
-#' @param nStrata2 Numeric. Number of quantiles to stratify \code{mraster2}.
-#' @param samp Numeric. Determines proportion of cells to plot in scatterplot (default = \code{0.2}).
-#' Lower values reduce visualization time.
+#' @param nStrata Numeric. A positive integer representing the number of desired classes 
+#' or a numeric vector of probabilities with values between \code{0-1}. If \code{mraster} has multiple layers,
+#' \code{nStrata} must be a list with an equal number of objects. 
 #'
 #' @return Returns an output stratification \code{spatRaster} or a list when \code{details = TRUE}.
 #'
 #' When a list is returned:
 #' \enumerate{
-#' \item \code{details} generates is a strata lookUp table for the stratification.
+#' \item \code{details} lookUp table for stratification(s).
 #' \item \code{raster} is a stratified \code{spatRaster} based on quantiles
 #' \item \code{plot} is a \code{ggplot} histogram / scatter plot object (depends on whether metric2 was supplied).
-#' Histogram shows distribution and break points while scatter plot shows colour coded and strata boundaries.
+#' Histogram shows distribution and break points.
 #' }
 #'
 #' @examples
@@ -27,269 +26,164 @@
 #'
 #' strat_quantiles(
 #'   mraster = mr$zq90,
-#'   nStrata = 4,
-#'   plot = TRUE
+#'   nStrata = 4
 #' )
 #'
 #' strat_quantiles(
-#'   mraster = mr$zq90,
-#'   mraster2 = mr$zsd,
-#'   nStrata = 3,
-#'   nStrata2 = 4
+#'   mraster = mr[[1:2]],
+#'   nStrata = list(c(0.2,0.4,0.8), 3),
+#'   map = TRUE
 #' )
+#'   
 #' @author Tristan R.H. Goodbody
 #'
 #' @export
 
 strat_quantiles <- function(mraster,
-                            mraster2 = NULL,
                             nStrata,
-                            nStrata2 = NULL,
+                            map = FALSE,
                             plot = FALSE,
                             details = FALSE,
-                            samp = 0.2,
                             filename = NULL,
                             overwrite = FALSE) {
 
   #--- Set global vars ---#
-
-  class1 <- class2 <- NULL
+  strata <- x <- y <- value <- val <- NULL
 
   #--- error handling ---#
   if (!inherits(mraster, "SpatRaster")) {
     stop("'mraster' must be type SpatRaster.", call. = FALSE)
   }
-
-  if (!is.numeric(nStrata)) {
-    stop("'nStrata' must be type numeric.", call. = FALSE)
+  
+  if (!is.logical(map)) {
+    stop("'map' must be type logical.", call. = FALSE)
   }
 
   if (!is.logical(plot)) {
     stop("'plot' must be type logical.", call. = FALSE)
   }
 
-  if (!is.numeric(samp)) {
-    stop("'samp' must be type numeric.", call. = FALSE)
-  }
-
   if (!is.logical(details)) {
     stop("'details' must be type logical.", call. = FALSE)
   }
-
+  
+  #--- get number of layers ---#
+  nlayer <- terra::nlyr(mraster)
+  
   #--- if there is only 1 band in mraster use it as default ---#
-
-  if (terra::nlyr(mraster) > 1) {
-    stop("Multiple layers detected in 'mraster'. Please define a singular band to stratify.", call. = FALSE)
+  if(nlayer > 1){
+    
+    if (nlayer != length(nStrata)) {
+      stop("`mraster` and `nStrata` must have the same number of layers & objects.", call. = FALSE)
+    }
+    
+    if(!inherits(nStrata, "list")){
+      stop("`nStrata` must be a list of numeric vectors of the same length as `mraster`.", call. = FALSE)
+    }
+    
+  } else{
+    
+    if(!is.list(nStrata)){
+      nStrata <- list(nStrata)
+    }
+    
   }
-
-  if (!is.null(mraster2)) {
-    if (!inherits(mraster2, "SpatRaster")) {
-      stop("'mraster2' must be type SpatRaster.", call. = FALSE)
-    }
-
-    if (isFALSE(terra::compareGeom(mraster, mraster2, stopOnError = FALSE))) {
-      stop("Extents of 'mraster' and 'mraster2' do not match.", call. = FALSE)
-    }
+  
+  #--- convert mraster to list ---#
+  mrl <- as.list(mraster)
+  
+  #--- vectorize vect_breaks to determine raster break points ---#
+  rastbreaks <- mapply(calculate_quantile_breaks, mraster = mrl, nStrata = nStrata)
+  
+  #--- take all raster outputs from vectorization ---#
+  rstack <- terra::rast(rastbreaks[c(TRUE, FALSE)])
+  
+  #--- rename to append raster metric name ---#
+  names(rstack) <- paste0("strata_",names(mraster))
+  
+  if(nlayer > 1){
+    #--- establish newly formed unique class ---#
+    breaks_c <- terra::as.data.frame(rstack, xy = TRUE)%>%
+      dplyr::group_by(dplyr::across(tidyr::starts_with("strata"))) %>%
+      dplyr::mutate(strata = dplyr::cur_group_id()) %>%
+      dplyr::ungroup()
     
-    if (isFALSE(terra::compareGeom(mraster, mraster2, stopOnError = FALSE, ext = FALSE, res = TRUE))) {
-      stop("Spatial resolutions of 'mraster' and 'mraster2' do not match.", call. = FALSE)
-    }
-
-    #--- if there is only 1 band in mraster2 use it as default ---#
-
-    if (terra::nlyr(mraster2) > 1) {
-      stop("Multiple layers detected in 'mraster2'. Please define a singular band to stratify.", call. = FALSE)
-    }
-
-    #--- ensure names of rasters are not the same ---#
-
-    if (identical(names(mraster), names(mraster2))) {
-      stop("'mraster' and 'mraster2' must have different metric names.", call. = FALSE)
-    }
+    lookUp <- dplyr::arrange(breaks_c, strata) %>% 
+      dplyr::select(-x,-y) %>% 
+      unique()
     
-    if (!is.numeric(nStrata2)) {
-      stop("'nStrata2' must be type numeric.", call. = FALSE)
-    }
-  }
-
-  if (is.null(mraster2)) {
-    vals <- terra::values(mraster)
-
-    vals[!is.finite(vals)] <- NA
-
-    #--- Determine index of each cell so to map values correctly without NA ---#
-
-    idx <- !is.na(vals)
-
-    #--- Remove NA / NaN / Inf values ---#
-
-    df <- vals %>%
-      as.data.frame() %>%
-      stats::na.omit()
-
-    metric <- as.character(names(df))
-
-    metric <- ggplot2::ensym(metric)
-
-    #--- Split mraster distribution in to number specified by 'breaks' ---#
-
-    dfc <- df %>%
-      dplyr::mutate(class = dplyr::ntile(!!metric, nStrata))
+    idx <- terra::cellFromXY(mraster[[1]], cbind(breaks_c$x, breaks_c$y))
     
-    #--- generate lookup table for details output ---#
-    lookUp <- dfc %>%
-      dplyr::group_by(class) %>%
-      dplyr::summarize(min_mraster = min(!!metric),
-                    max_mraster = max(!!metric))
-
-    #--- convert back to original mraster extent ---#
-
-    odf <- matrix(nrow = nrow(vals), ncol = 1)
+    rcl <- mraster[[1]]
+    #--- convert back to original extent ---#
+    rcl[idx] <- breaks_c$strata
     
-    odf[, 1][idx] <- dfc$class
-
-    #--- set newly stratified values ---#
-
-    rout <- terra::setValues(mraster[[1]], odf[, 1])
-    names(rout) <- "strata"
-
-    #--- if mraster2 is provided ---#
-  } else {
-
-    #--- Extract values from mraster ---#
-
-    vals <- c(mraster, mraster2) %>%
-      terra::values()
-
-    vals[!is.finite(vals)] <- NA
-
-    #--- Determine index of each cell so to map values correctly without NA ---#
-
-    idx <- is.finite(vals[, 1]) & is.finite(vals[, 2])
-
-    #--- Remove NA / NaN / Inf values ---#
-
-    df <- vals %>%
-      as.data.frame() %>%
-      stats::na.omit()
-
-    nm <- as.character(names(df))
-
-    metric <- nm[1]
-    metric2 <- nm[2]
-
-    metric <- ggplot2::ensym(metric)
-    metric2 <- ggplot2::ensym(metric2)
-
-    #--- Split metric distribution in to number specified by 'breaks' ---#
-
-    dfc <- df %>%
-      #--- define nStrata classes ---#
-      dplyr::mutate(class1 = dplyr::ntile(!!metric, nStrata)) %>%
-      #--- group by class to sub stratify ---#
-      dplyr::group_by(class1) %>%
-      #--- define nStrata2 classes ---#
-      dplyr::mutate(class2 = dplyr::ntile(!!metric2, nStrata2)) %>%
-      #--- combine classes ---#
-      dplyr::group_by(class1, class2) %>%
-      #--- establish newly formed unique class ---#
-      dplyr::mutate(class = dplyr::cur_group_id())
+    names(rcl) <- "strata"
     
-    #--- generate lookup table for details output ---#
-    lookUp <- dfc %>%
-      dplyr::group_by(class) %>%
-      dplyr::summarize(min_mraster = min(!!metric),
-                       max_mraster = max(!!metric),
-                       min_mraster2 = min(!!metric2),
-                       max_mraster2 = max(!!metric2))
-
-    #--- convert back to original mraster extent ---#
-
-    odf <- matrix(nrow = nrow(vals), ncol = 1)
-    
-    odf[, 1][idx] <- dfc$class
-
-    #--- set newly stratified values ---#
-
-    rout <- terra::setValues(mraster[[1]], odf[, 1])
-    names(rout) <- "strata"
-  }
-
-  if (isTRUE(plot)) {
-    if (is.null(mraster2)) {
-
-      #--- output histogram with quantile breaks ---#
-
-      #--- determine numeric breaks ---#
-
-      breaks <- dfc %>%
-        dplyr::group_by(class) %>%
-        dplyr::summarize(breaks = max(!!metric)) %>%
-        dplyr::select(breaks) %>%
-        as.data.frame()
-
-      breaks <- breaks[1:(nStrata - 1), ]
-
-      df.p <- dfc %>%
-        dplyr::select(!!metric)
-
-      #--- plot histogram of metric with associated break lines ---#
-
-      p <- ggplot2::ggplot(df.p, ggplot2::aes(!!metric)) +
-        ggplot2::geom_histogram() +
-        ggplot2::geom_vline(xintercept = breaks, linetype = "dashed") +
-        ggplot2::ggtitle(paste0(metric, " histogram with defined breaks"))
-
-      terra::plot(rout, main = "Classes")
+    if(isTRUE(map)){
+      
+      message("Mapping stratifications.")
+      
+      rcl <- c(rstack,rcl)
+      
     } else {
-      terra::plot(rout, main = "Classes")
+      
+      rcl <- rstack
+      
+      names(rcl) <- rep("strata",nlayer)
+      
     }
+    
+  } else {
+    
+    rcl <- rstack
+    
+    names(rcl) <- "strata"
+  }
+  
+  if (isTRUE(plot)) {
+    
+    #--- take numeric breaks for plotting ---#
+    brs <- do.call("rbind",rastbreaks[c(FALSE, TRUE)])
+
+    brs <- brs[is.finite(brs$val),]
+
+    p <- terra::as.data.frame(mraster) %>%
+      tidyr::pivot_longer(dplyr::everything(),names_to = "names") %>%
+      ggplot2::ggplot(ggplot2::aes(value)) +
+      ggplot2::geom_histogram() +
+      ggplot2::geom_vline(linetype = "dashed", data = brs, mapping = ggplot2::aes(xintercept = val)) +
+      ggplot2::facet_wrap(~names, scales = "free")
+    
+    suppressMessages(print(p))
+    
+    #--- set colour palette ---#
+    
+    terra::plot(rcl)
   }
 
   #--- write file to disc ---#
 
   if (!is.null(filename)) {
-    terra::writeRaster(x = rout, filename = filename, overwrite = overwrite)
+    terra::writeRaster(x = rcl, filename = filename, overwrite = overwrite)
     message("Output raster written to disc.")
   }
 
   #--- Output based on 'details' to return raster alone or list with details ---#
 
   if (isTRUE(details)) {
-    if (!is.null(mraster2)) {
-      if (samp > 1 | samp < 0) {
-        stop("'samp' must be > 0 <= 1.", call. = FALSE)
-      }
-
-      #--- create classplot summary ---#
-
-      coordsgrps <- dfc %>%
-        dplyr::group_by(class) %>%
-        dplyr::arrange(class) %>%
-        stats::na.omit() %>%
-        tidyr::nest() %>%
-        dplyr::ungroup()
-
-      p <- classPlot(as.data.frame(dfc),
-        coordsgrps = coordsgrps,
-        mraster = !!metric,
-        mraster2 = !!metric2,
-        samp
-      )
-    }
-
     #--- output metrics details along with stratification raster ---#
 
     if (isTRUE(plot)){
     out <- list(
       details = lookUp,
-      raster = rout,
+      raster = rcl,
       plot = p
     )
     } else {
       out <- list(
         details = lookUp,
-        raster = rout
+        raster = rcl
       )
     }
 
@@ -298,6 +192,6 @@ strat_quantiles <- function(mraster,
 
     #--- just output raster ---#
 
-    return(rout)
+    return(rcl)
   }
 }