+++ tests/testthat/test-sample_existing.R
@@ -33,7 +33,7 @@ test_that("categorical", {
 
 
 test_that("Access", {
-  expect_message(sample_existing(existing = existing, raster = mraster, nSamp = 20, access = access, buff_inner = 50, buff_outer = 200, plot = TRUE), "An access layer has been provided. An internal buffer of 50 m and an external buffer of 200 m have been applied.")
+  expect_message(sample_existing(existing = existing, raster = mraster, nSamp = 20, access = access, buff_inner = 50, buff_outer = 200), "An access layer has been provided. An internal buffer of 50 m and an external buffer of 200 m have been applied.")
   expect_message(sample_existing(existing = existing, raster = mraster, nSamp = 20, access = access, buff_outer = 200), "An access layer has been provided. An external buffer of 200 m have been applied.")
   expect_message(sample_existing(existing = existing, raster = mraster, nSamp = 20, filename = file.path(tempdir(), "temp.shp"), overwrite = TRUE), "Output samples written to disc.")
   expect_message(sample_existing(raster = mraster, nSamp = 1, existing = existingna), "16 samples are located where metric values are NA.")